[2/N][torch.compile] make compilation cfg part of vllm cfg (#10383)
